+++ lib/glimmer/swt/custom/animation.rb
@@ -215,11 +215,8 @@ module Glimmer
           self.class.schedule_frame_animation(self) do
             if started? && start_number == @start_number && within_duration_limit?
               unless @parent.isDisposed
-                @parent.clear_shapes
-                @parent.content {
-                  frame_block.call(*block_args)
-                }
-                @parent.redraw
+                @parent.clear_shapes # TODO adjust this to clear only the shapes of this animation (not all shapes) to allow simultaneous animations to occur on the same parent
+                @parent.content { frame_block.call(*block_args) }
               end
             else
               if stopped? && @frame_index > current_frame_index